move number range
