+++ src/packages/core/property-editor/uis/manifests.ts
@@ -8,7 +8,7 @@ import { manifest as iconPicker } from '../../../property-editors/icon-picker/ma
 import { manifest as label } from '../../../property-editors/label/manifests.js';
 import { manifest as multipleTextString } from '../../../property-editors/multiple-text-string/manifests.js';
 import { manifest as multiUrlPicker } from '../../../property-editors/multi-url-picker/manifests.js';
-import { manifest as numberRange } from './number-range/manifests.js';
+import { manifest as numberRange } from '../../../property-editors/number-range/manifests.js';
 import { manifest as orderDirection } from './order-direction/manifests.js';
 import { manifest as overlaySize } from './overlay-size/manifests.js';
 import { manifest as radioButtonList } from './radio-button-list/manifests.js';
