+++ src/packages/property-editors/number-range/manifests.ts
